Move #dummy_tx to Glueby::Contract::FeeEstimator
